pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "earthquake-olap-showcase"
version = "0.1.0"
description = "OLAP analytics showcase using DuckDB, Python, and Streamlit with earthquake data"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["duckdb", "olap", "data-engineering", "streamlit", "earthquake", "analytics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "duckdb>=0.9.2",
    "streamlit>=1.28.0",
    "polars>=0.19.0",
    "pandas>=2.1.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "plotly>=5.17.0",
    "folium>=0.15.0",
    "streamlit-folium>=0.15.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "rich>=13.6.0",
    "skyfield>=1.48",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.9.0",
    "ipython>=8.16.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/dtc77ai/earthquake-olap-showcase"
Repository = "https://github.com/dtc77ai/earthquake-olap-showcase"
Issues = "https://github.com/dtc77ai/earthquake-olap-showcase/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=60",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/app/main.py",  # Streamlit app is hard to test
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
